--- src/pages/About.jsx
@@ -1,5 +1,0 @@
-function About() {
-  return <div>About</div>;
-}
-
-export default About;
