+++ src/pages/About/About.jsx
@@ -1,0 +1,5 @@
+function About() {
+  return <div>About</div>;
+}
+
+export default About;
