--- src/pages/Contact.jsx
@@ -1,5 +1,0 @@
-function Contact() {
-  return <div>Contact</div>;
-}
-
-export default Contact;
