+++ src/pages/Contact/Contact.jsx
@@ -1,0 +1,5 @@
+function Contact() {
+  return <div>Contact</div>;
+}
+
+export default Contact;
