--- src/pages/Home.jsx
@@ -1,5 +1,0 @@
-function Home() {
-  return <div>Home</div>;
-}
-
-export default Home;
