--- src/pages/SignUp.jsx
@@ -1,5 +1,0 @@
-function SignUp() {
-  return <div>SignUp</div>;
-}
-
-export default SignUp;
