+++ src/pages/Signup/SignUp.jsx
@@ -1,0 +1,5 @@
+function SignUp() {
+  return <div>SignUp</div>;
+}
+
+export default SignUp;
