+++ src/routes.js
@@ -1,4 +1,4 @@
-export const HOME_ROUTE = '/home';
+export const HOME_ROUTE = '/';
 export const CONTACT_ROUTE = '/contact';
 export const ABOUT_ROUTE = '/about';
 export const LOGIN_ROUTE = '/login';
